Moved Mutation definition to separate file with companion object
